chore(validator): extract fixtures into a separate folder
