+++ packages/validator/tests/fixtures/oas2.github.json
