+++ packages/validator/tests/fixtures/oas2.petstore.json
