+++ packages/validator/tests/fixtures/oas3.petstore.yaml
